put TorState variable in new BaseEventBroadcaster class and move EventBroadcaster to topl-core-base module
