+++ topl-core/src/main/java/io/matthewnelson/topl_core/broadcaster/TorStateMachine.kt
@@ -12,6 +12,7 @@ See the Apache 2 License for the specific language governing permissions and lim
 */
 package io.matthewnelson.topl_core.broadcaster
 
+import io.matthewnelson.topl_core_base.EventBroadcaster
 import io.matthewnelson.topl_core_base.TorStates
 
 /**
